Move useBlockBlobTxsQuery to data-availability/hooks/
